python/requirements.txt
google-cloud-bigquery>=3.22.0
